Add core/existing RobotStanvern functionality.
This is a slightly-modified version of the existing code that has been running for a ~year (changes: secrets moved to config file, small cleanups done, some extra stuff removed). Although it compiles, beyond some basic checks, I haven't actually run it to verify the full functionality works as intended. I'll be switching the version I actually run to this and weeding out any potential small issues, but in the meantime, this should suffice as an example of roughly how to put something like this together.