Remove the ability for the FTE listener to play a local sound.
The only point of this was to be able to idle in a raid zone and get notified of FTEs while doing other stuff, sleeping, etc. Now that the batphone relay is integrated with PagerDuty, this isn't needed. Also, the long sound clip being played on every FTE was annoying anyway in the case of multiple FTEs within a short period of time.

This change also removes the constant that lets Discord messages on FTEs be disabled (the constant that makes it be sent as a DM remains).