Add a Discord listener that replies to screenshots of the char info part of the UI with a scrape summary of name/class/level/exp.
The core bits of logic are (a) Tesseract for reading the text and (b) a pixel grabber for reading the exp bar. Given the fuzzy/imperfect nature of both of these and the varying backgrounds on which the relevant part of the UI can be on top of, it's likely that this logic will need to be extended over time. Unit tests are in place to ensure that no regressions creep in when modifying the logic.