Make game client screenshotting more robust.
If the game client window is on a different desktop than what's currently in use OR if it's minimized, then the screenshot will be stale (i.e. it'll be whatever was last rendered when the window was visible). Avoid this problem by activating the window, which will switch to its desktop and un-minimize if needed. After that, re-minimize the game client window and switch back to the original desktop as needed.