Look for item names only in actual guild chat message (i.e. not in the "$player tells the guild" part).
If the player name happens to contain an item name, then this will otherwise result in an unwanted item name match.

Note that this doesn't solve the problem when the name appears in the actual grats message (which will often be the same as the author), but I'm still adding this anyway because (a) it's incorrect to look outside the actual /gu message for parsing and (b) I don't have a perfect solution to the underlying problem, so this would still help in the case where author != winner.