Use acknowledge() instead of respond() for grats channel button clicks.
At some point, clicking the button started showing a "This interaction failed" error (benign - everything still worked as intended). The cause was an uncaught exception in the response from Discord on respond() - apparently that method is meant for when the button interaction is going to be replied to. My best guess is that some Discord serverside validation update was introduced at some point to enforce this. In any case, since the bot doesn't need to actually respond but rather just acknowledge, the fix makes sense.