Only attempt to process screenshots from past 24 hours.
For some reason, there are currently a couple months worth of screenshots that haven't been processed. Because the processor scans for unprocessed screenshots on startup, and spins up a new expensive processor for each screenshot in parallel, this ended up choking the server after restarting the Discord bot. Capping the lookback window to 24 hours mitigates this. In practice, the Discord bot has never been down that long - plus it's not a big deal for these screenshots to not get automatically processed.